fix(ui): keep bare dollar math disabled
